Posting data
With these changes, tried on a new DB, all fields are correctly posted except the computed bmi value. A value of 0 is sent. I think the problem is on the bmi state updating. It seems to update on the second submit, even with new data, which should not be. It is supposed to be submitted along with the other data fields. I am going to work on this.